added plot, datamodel example in readme, renamed spec.py to datamodel.py
