+++ README.md
@@ -203,6 +203,51 @@ qsoabsfind --config example_config.yaml --absorber CIV --verbose
 
 A fully annotated template is provided at `data/example_config.yaml`.
 
+**Reading output catalogs**
+---------------------
+
+After running the absorber search, you can load the output FITS catalog using the `AbsorberData` class:
+
+```python
+from qsoabsfind.datamodel import AbsorberData
+
+catalog = AbsorberData('test_MgII.fits', autoload=True)
+
+print(catalog.catalog)        # absorber table (ABSORBER HDU)
+print(catalog.metadata)       # QSO metadata (METADATA HDU)
+print(catalog.column_density) # column densities if present, else None
+```
+
+**Plotting a random absorber**
+---------------------
+
+Once you have loaded the spectra and the output catalog, you can visualise a randomly selected absorber using `plot_absorber` from `qsoabsfind.utils`:
+
+```python
+import numpy as np
+from qsoabsfind.datamodel import QSOSpecRead, AbsorberData
+from qsoabsfind.utils import plot_absorber
+
+# Load the output absorber catalog
+catalog = AbsorberData('/path/to/your/absorber.fits', autoload=True)
+
+# Pick a random absorber from the catalog
+rng = np.random.default_rng()
+idx = rng.integers(len(catalog.catalog))
+row = catalog.catalog[idx]
+
+# Load the corresponding QSO spectrum
+spectra = QSOSpecRead('/path/to/your/spectra.fits',
+                      index=int(row['INDEX_SPEC']),
+                      autoload=True)
+
+# Plot the absorber (full spectrum + zoomed-in doublet view)
+plot_absorber(spectra, absorber='MgII', zabs=row,
+              title=f"MgII absorber at z={row['Z_ABS']:.4f}")
+```
+
+Pass `show_error=True` to overlay the error spectrum, or `plot_filename='absorber.png'` to save the figure to disk instead of displaying it interactively.
+
 **Pre-filtering searchable QSOs**
 ---------------------
 
@@ -245,7 +290,7 @@ Example catalog runs
 
 SDSS and DESI [example jupyter notebooks](https://github.com/abhi0395/qsoabsfind/blob/main/nb/) are also available.
 
-Citation & Acknowledgements
+Citation 
 ---------------------------
 
 If you use this code in your analysis, please cite [Anand, Nelson & Kauffmann 2021](https://arxiv.org/abs/2103.15842) and [Anand et al. 2025](https://arxiv.org/abs/2504.20299). The BibTeX entries for these papers can be found [here (2021 paper)](https://ui.adsabs.harvard.edu/abs/2021MNRAS.504...65A/exportcitation) and [here (2025 paper)](https://ui.adsabs.harvard.edu/abs/2025arXiv250420299A/exportcitation).
@@ -314,7 +359,7 @@ Contributions are welcome! Please submit a pull request or open an issue to disc
 Acknowledgements
 -----------
 
-The first crude version of the codebase was developed and written by me during my PhD with lots of suggestions from my PhD supervisors [Prof. Dr. Guinevere Kauffmann](https://www.mpa-garching.mpg.de/person/44092) and [Dr. Dylan Nelson](https://nelson.tng-project.org/). Over the years, it has evolved from a specialized script into the generic, community-ready framework it is today. I would like to extend my thanks to the VS Code AI agents, which were instrumental in refining the codebase. They provided invaluable assistance in documenting functions, optimizing logic, and expanding unit test coverage. They helped ensure the code is both robust and maintainable. The project logo was created from a absorber example generated by me, with assistance from ChatGPT-5.
+The first crude version of the codebase was developed and written by me during my PhD with lots of suggestions from my PhD supervisors [Prof. Dr. Guinevere Kauffmann](https://www.mpa-garching.mpg.de/person/44092) and [Dr. Dylan Nelson](https://nelson.tng-project.org/). Over the years, it has evolved from a specialized script into the generic, community-ready framework it is today. I would like to extend my thanks to the VS Code AI agents, which were instrumental in refining the codebase. They provided invaluable assistance in documenting functions, loggers, optimizing logic, and expanding unit test coverage. They helped ensure the code is both robust and maintainable. The project logo was created from a absorber example generated by me, with assistance from ChatGPT-5.
 
 License
 -------
